Moved BootDriver to Common area
