+++ BoardSupport/qemu-aarch64/Boot/CMakeLists.txt
@@ -1,6 +1,3 @@
-LIST(APPEND BOOT_HEADERS BootDriver.hpp)
-LIST(APPEND BOOT_SOURCES BootDriver.cpp)
-
 LIST(APPEND BOOT_SOURCES Boot.cpp)
 
 ADD_LIBRARY(Boot STATIC ${BOOT_HEADERS} ${BOOT_SOURCES})
