+++ BoardSupport/qemu-aarch64/CMakeLists.txt
@@ -1,6 +1,9 @@
 LIST(APPEND BOARD_HEADERS Board.hpp)
 LIST(APPEND BOARD_SOURCES Board.cpp)
 
+LIST(APPEND BOARD_HEADERS ../Common/Boot/BootDriver.hpp)
+LIST(APPEND BOARD_SOURCES ../Common/Boot/BootDriver.cpp)
+
 LIST(APPEND BOARD_HEADERS ../Common/DeviceTree/DeviceTree.hpp)
 LIST(APPEND BOARD_SOURCES ../Common/DeviceTree/DeviceTree.cpp)
